refactor: move QuoteError definition and implementation to solver-types module
