+++ crates/solver-types/src/api.rs
@@ -195,7 +195,6 @@ impl fmt::Display for APIError {
 
 impl std::error::Error for APIError {}
 
-#[cfg(feature = "axum")]
 impl axum::response::IntoResponse for APIError {
 	fn into_response(self) -> axum::response::Response {
 		use axum::{http::StatusCode, response::Json};
@@ -234,3 +233,48 @@ pub mod u256_serde {
 	}
 }
 
+/// Errors that can occur during quote processing.
+#[derive(Debug, thiserror::Error)]
+pub enum QuoteError {
+    #[error("Invalid request: {0}")]
+    InvalidRequest(String),
+    #[error("Unsupported asset: {0}")]
+    UnsupportedAsset(String),
+    #[error("Insufficient liquidity for requested amount")]
+    InsufficientLiquidity,
+    #[error("Solver capacity exceeded")]
+    SolverCapacityExceeded,
+    #[error("Internal error: {0}")]
+    Internal(String),
+}
+
+impl From<QuoteError> for APIError {
+    fn from(quote_error: QuoteError) -> Self {
+        match quote_error {
+            QuoteError::InvalidRequest(msg) => APIError::BadRequest {
+                error_type: "INVALID_REQUEST".to_string(),
+                message: msg,
+                details: None,
+            },
+            QuoteError::UnsupportedAsset(asset) => APIError::UnprocessableEntity {
+                error_type: "UNSUPPORTED_ASSET".to_string(),
+                message: format!("Asset not supported by solver: {}", asset),
+                details: Some(serde_json::json!({ "asset": asset })),
+            },
+            QuoteError::InsufficientLiquidity => APIError::UnprocessableEntity {
+                error_type: "INSUFFICIENT_LIQUIDITY".to_string(),
+                message: "Insufficient liquidity available for the requested amount".to_string(),
+                details: None,
+            },
+            QuoteError::SolverCapacityExceeded => APIError::ServiceUnavailable {
+                error_type: "SOLVER_CAPACITY_EXCEEDED".to_string(),
+                message: "Solver capacity exceeded, please try again later".to_string(),
+                retry_after: Some(60), // Suggest retry after 60 seconds
+            },
+            QuoteError::Internal(msg) => APIError::InternalServerError {
+                error_type: "INTERNAL_ERROR".to_string(),
+                message: format!("An internal error occurred: {}", msg),
+            },
+        }
+    }
+}
